Revert "Added replaceInString & getConfig (#86)"
This reverts commit 1ff7cae4002a6824422f4a38f1b06d9de916793f.
